[REF] im_livechat,website_livechat: move fields to channel
