+++ addons/crm_livechat/static/src/core/web/livechat_channel_info_list_patch.xml
@@ -2,7 +2,7 @@
 <templates xml:space="preserve">
     <t t-name="website_livechat.LivechatChannelInfoList" t-inherit="im_livechat.LivechatChannelInfoList" t-inherit-mode="extension">
         <xpath expr="//t[@t-name='extra_infos']" position="inside">
-            <t t-if="this.props.thread.livechatVisitorMember?.partner_id?.opportunity_ids?.length" t-call="im_livechat.LivechatChannelInfoList.info_links">
+            <t t-if="this.props.thread.channel.livechatVisitorMember?.partner_id?.opportunity_ids?.length" t-call="im_livechat.LivechatChannelInfoList.info_links">
                 <t t-set="title">Open leads</t>
                 <t t-set="templateParams"
                     t-value="{
