+++ addons/im_livechat/static/src/core/common/thread_patch.js
@@ -1,8 +1,8 @@
-import { useLayoutEffect } from "@web/owl2/utils";
 import { Thread } from "@mail/core/common/thread";
 import { _t } from "@web/core/l10n/translation";
 import { user } from "@web/core/user";
 import { patch } from "@web/core/utils/patch";
+import { useLayoutEffect } from "@web/owl2/utils";
 
 const { DateTime } = luxon;
 
@@ -10,14 +10,14 @@ patch(Thread.prototype, {
     setup() {
         super.setup(...arguments);
         this.IM_STATUS_DELAY = 1500;
-        Object.assign(this.state, { isVisitorOffline: false }); // starting online avoids flickering
+        this.state.isVisitorOffline = false; // starting online avoids flickering
         useLayoutEffect(
-            () => {
-                if (!this.props.thread.livechatVisitorMember?.im_status) {
+            (im_status) => {
+                if (!im_status) {
                     return;
                 }
                 clearTimeout(this.imStatusTimeoutId);
-                if (this.props.thread.livechatVisitorMember.im_status === "offline") {
+                if (im_status === "offline") {
                     this.imStatusTimeoutId = setTimeout(
                         () => (this.state.isVisitorOffline = true),
                         this.IM_STATUS_DELAY
@@ -27,19 +27,19 @@ patch(Thread.prototype, {
                 }
                 return () => clearTimeout(this.imStatusTimeoutId);
             },
-            () => [this.props.thread.livechatVisitorMember?.im_status]
+            () => [this.props.thread.channel?.livechatVisitorMember?.im_status]
         );
     },
     get showVisitorDisconnected() {
         return (
-            this.store.self.notEq(this.props.thread.livechatVisitorMember?.persona) &&
-            !this.props.thread.livechat_end_dt &&
-            this.props.thread.livechatVisitorMember &&
+            this.store.self.notEq(this.channel?.livechatVisitorMember?.persona) &&
+            !this.channel?.livechat_end_dt &&
+            this.channel?.livechatVisitorMember &&
             this.state.isVisitorOffline
         );
     },
     get disconnectedText() {
-        const offlineSince = this.props.thread.livechatVisitorMember.persona.offline_since;
+        const offlineSince = this.props.thread.channel?.livechatVisitorMember.persona.offline_since;
         if (!offlineSince) {
             return _t("Visitor is disconnected");
         }
