+++ addons/im_livechat/static/src/core/public_web/discuss_channel_model_patch.js
@@ -44,6 +44,13 @@ const discussChannelPatch = {
         );
     },
 
+    get inChathubOnNewMessage() {
+        if (this.channel_type === "livechat") {
+            return Boolean(this.channel.self_member_id);
+        }
+        return super.inChathubOnNewMessage;
+    },
+
     get livechatStatusLabel() {
         if (this.livechat_end_dt) {
             return _t("Conversation has ended");
