+++ addons/im_livechat/static/src/core/public_web/thread_model_patch.js
@@ -3,13 +3,6 @@ import { Thread } from "@mail/core/common/thread_model";
 import { patch } from "@web/core/utils/patch";
 
 patch(Thread.prototype, {
-    get inChathubOnNewMessage() {
-        if (this.channel?.channel_type === "livechat") {
-            return Boolean(this.channel.self_member_id);
-        }
-        return super.inChathubOnNewMessage;
-    },
-
     /**
      * @override
      * @param {boolean} pushState
