+++ addons/im_livechat/static/src/core/web/livechat_channel_info_list.js
@@ -44,8 +44,11 @@ export class LivechatChannelInfoList extends Component {
     }
 
     onBlurNote() {
-        prettifyMessageContent(this.props.thread.livechatNoteText).then((note) => {
-            rpc("/im_livechat/session/update_note", { channel_id: this.props.thread.id, note });
+        prettifyMessageContent(this.props.thread.channel.livechatNoteText).then((note) => {
+            rpc("/im_livechat/session/update_note", {
+                channel_id: this.props.thread.channel.id,
+                note,
+            });
         });
     }
 
@@ -59,14 +62,14 @@ export class LivechatChannelInfoList extends Component {
         this.actionService.doAction({
             type: "ir.actions.act_window",
             res_model: "res.partner",
-            res_id: this.props.thread.livechatVisitorMember.partner_id.id,
+            res_id: this.props.thread.channel.livechatVisitorMember.partner_id.id,
             views: [[false, "form"]],
             target: "current",
         });
     }
 
     get visitorProfileURL() {
-        const visitorMember = this.props.thread?.livechatVisitorMember;
+        const visitorMember = this.props.thread?.channel?.livechatVisitorMember;
         if (visitorMember?.partner_id) {
             return url(`/${startUrl()}/res.partner/${visitorMember.partner_id.id}`);
         }
