+++ addons/im_livechat/static/src/embed/common/discuss_channel_model_patch.js
@@ -23,6 +23,21 @@ patch(DiscussChannel, discussChannelStaticPatch);
 const discussChannelPatch = {
     setup() {
         super.setup(...arguments);
+        this.livechatWelcomeMessage = fields.One("mail.message", {
+            compute() {
+                if (this.hasWelcomeMessage) {
+                    const livechatService = this.store.env.services["im_livechat.livechat"];
+                    return {
+                        id: -0.2 - this.id,
+                        body: livechatService.options.default_message,
+                        thread: this.thread,
+                        author_id: this.livechat_agent_history_ids.sort((a, b) => a.id - b.id)[0]
+                            ?.partner_id,
+                    };
+                }
+            },
+        });
+        this.requested_by_operator = false;
         this.storeAsActiveLivechats = fields.One("Store", {
             compute() {
                 return this.channel_type === "livechat" && !this.livechat_end_dt
@@ -30,6 +45,26 @@ const discussChannelPatch = {
                     : null;
             },
             inverse: "activeLivechats",
+        });
+        this._toggleChatbot = fields.Attr(false, {
+            compute() {
+                return Boolean(
+                    this.channel?.chatbot &&
+                        !this.channel.chatbot.completed &&
+                        !this.channel.livechat_end_dt
+                );
+            },
+            onUpdate() {
+                const shouldToggle = this._toggleChatbot;
+                this.isLoadedDeferred.then(() => {
+                    if (shouldToggle) {
+                        this.channel.chatbot.start();
+                    } else {
+                        this.channel?.chatbot?.stop();
+                    }
+                });
+            },
+            eager: true,
         });
     },
     get avatarUrl() {
@@ -57,6 +92,12 @@ const discussChannelPatch = {
     },
     get hasAttachmentPanel() {
         return this.channel_type !== "livechat" && super.hasAttachmentPanel;
+    },
+    get hasWelcomeMessage() {
+        return this.channel_type === "livechat" && !this.chatbot && !this.requested_by_operator;
+    },
+    get isLastMessageFromCustomer() {
+        return this.newestPersistentOfAllMessage?.isSelfAuthored;
     },
     _onDeleteChatWindow() {
         if (this.isTransient && this.channel_type === "livechat") {
