+++ addons/im_livechat/static/src/embed/common/feedback_panel/feedback_panel.xml
@@ -28,7 +28,7 @@
         <div class="d-flex gap-2 justify-content-center">
             <button class="btn btn-outline-secondary" t-on-click="this.props.onClickClose">Close</button>
             <button t-if="this.allowNewSession" class="btn btn-outline-secondary" t-on-click="this.props.onClickNewSession">New Session</button>
-            <a t-if="this.store.can_download_transcript" class="btn btn-outline-secondary" title="Download a copy of this conversation" target="_blank" t-att-href="this.props.thread.transcriptUrl"><i class="fa fa-download"/></a>
+            <a t-if="this.store.can_download_transcript" class="btn btn-outline-secondary" title="Download a copy of this conversation" target="_blank" t-att-href="this.props.thread.channel?.transcriptUrl"><i class="fa fa-download"/></a>
         </div>
     </div>
 </div>
