+++ addons/im_livechat/static/src/embed/common/thread_patch.xml
@@ -2,8 +2,8 @@
 <templates xml:space="preserve">
     <t t-inherit="mail.Thread" t-inherit-mode="extension">
         <xpath expr="//*[@name='content']" position="before">
-            <div t-if="this.props.thread?.livechatWelcomeMessage" class="bg-100 py-3" data-name="LivechatWelcomeMessage">
-                <Message message="this.props.thread.livechatWelcomeMessage" hasActions="false" thread="this.props.thread"/>
+            <div t-if="this.props.thread?.channel?.livechatWelcomeMessage" class="bg-100 py-3" data-name="LivechatWelcomeMessage">
+                <Message message="this.props.thread.channel.livechatWelcomeMessage" hasActions="false" thread="this.props.thread"/>
             </div>
         </xpath>
         <xpath expr="(//div[hasclass('o-mail-Thread')])" position="inside">
