+++ addons/im_livechat/static/src/embed/cors/@types/models.d.ts
@@ -1,5 +1,5 @@
 declare module "models" {
-    export interface Thread {
+    export interface DiscussChannel {
         transcriptUrl: Readonly<string>;
     }
 }
