+++ addons/im_livechat/static/src/embed/cors/discuss_channel_model_patch.js
@@ -1,11 +1,11 @@
-import { Thread } from "@mail/core/common/thread_model";
+import { DiscussChannel } from "@mail/discuss/core/common/discuss_channel_model";
 import { expirableStorage } from "@im_livechat/core/common/expirable_storage";
 import { GUEST_TOKEN_STORAGE_KEY } from "@im_livechat/embed/common/store_service_patch";
 
 import { patch } from "@web/core/utils/patch";
 import { url } from "@web/core/utils/urls";
 
-patch(Thread.prototype, {
+patch(DiscussChannel.prototype, {
     get transcriptUrl() {
         const guestToken = expirableStorage.getItem(GUEST_TOKEN_STORAGE_KEY);
         return url(`/im_livechat/cors/download_transcript/${this.id}`, {
