+++ addons/mail/static/src/core/public_web/@types/models.d.ts
@@ -7,7 +7,6 @@ declare module "models" {
     export interface Thread {
         askLeaveConfirmation: (body: string) => Promise<void>;
         discussAppAsThread: DiscussApp;
-        inChathubOnNewMessage: Readonly<boolean>;
         setActiveURL: () => void;
         setAsDiscussThread: (pushState: boolean) => void;
     }
