+++ addons/website_livechat/static/src/core/web/livechat_channel_info_list_patch.xml
@@ -11,7 +11,7 @@
                 </div>
                 <span t-out="visitor.pageVisitHistoryText"/>
             </div>
-            <div t-if="!this.props.thread.livechat_end_dt and this.recentConversations.length" class="d-flex flex-column bg-inherit gap-1">
+            <div t-if="!this.props.thread.channel.livechat_end_dt and this.recentConversations.length" class="d-flex flex-column bg-inherit gap-1">
                 <h6 class="pt-3">Recent conversations</h6>
                 <div class="btn btn-group o-rounded-bubble d-flex flex-column w-100 p-0 m-0" style="gap: 1px;">
                     <a
